delete profile dropdown component
