+++ apps/web/src/contexts/globalLayout/GlobalNavbar/CollectionNavbar/CollectionNavbar.tsx
@@ -15,10 +15,6 @@ import {
   BreadcrumbLink,
   BreadcrumbText,
 } from '~/contexts/globalLayout/GlobalNavbar/ProfileDropdown/Breadcrumbs';
-import {
-  ProfileDropdown,
-  SlashText,
-} from '~/contexts/globalLayout/GlobalNavbar/ProfileDropdown/ProfileDropdown';
 import { CollectionNavbarFragment$key } from '~/generated/CollectionNavbarFragment.graphql';
 import { useIsMobileOrMobileLargeWindowWidth } from '~/hooks/useWindowSize';
 import unescape from '~/utils/unescape';
@@ -41,7 +37,7 @@ export function CollectionNavbar({ queryRef, username, collectionId }: Collectio
     graphql`
       fragment CollectionNavbarFragment on Query {
         ...CollectionRightContentFragment
-        ...ProfileDropdownFragment
+
         ...NavActionFollowQueryFragment
 
         userByUsername(username: $username) @required(action: THROW) {
@@ -91,30 +87,27 @@ export function CollectionNavbar({ queryRef, username, collectionId }: Collectio
   return (
     <StandardNavbarContainer>
       <NavbarLeftContent>
-        <ProfileDropdown
-          queryRef={query}
-          rightContent={
-            isMobile ? null : (
-              <RightContentWrapper shrink gap={4}>
-                {query.userByUsername && (
-                  <NavActionFollow userRef={query.userByUsername} queryRef={query} />
-                )}
+        <Wrapper gap={4} align="center">
+          {isMobile ? null : (
+            <RightContentWrapper shrink gap={4}>
+              {query.userByUsername && (
+                <NavActionFollow userRef={query.userByUsername} queryRef={query} />
+              )}
 
-                <SlashText>/</SlashText>
+              <SlashText>/</SlashText>
 
-                <Link href={galleryRoute} legacyBehavior>
-                  <BreadcrumbLink href={route(galleryRoute)}>{galleryName}</BreadcrumbLink>
-                </Link>
+              <Link href={galleryRoute} legacyBehavior>
+                <BreadcrumbLink href={route(galleryRoute)}>{galleryName}</BreadcrumbLink>
+              </Link>
 
-                <SlashText>/</SlashText>
+              <SlashText>/</SlashText>
 
-                <CollectionNameText title={unescapedCollectionName}>
-                  {unescapedCollectionName}
-                </CollectionNameText>
-              </RightContentWrapper>
-            )
-          }
-        />
+              <CollectionNameText title={unescapedCollectionName}>
+                {unescapedCollectionName}
+              </CollectionNameText>
+            </RightContentWrapper>
+          )}
+        </Wrapper>
       </NavbarLeftContent>
 
       <NavbarCenterContent>
@@ -136,6 +129,11 @@ export function CollectionNavbar({ queryRef, username, collectionId }: Collectio
     </StandardNavbarContainer>
   );
 }
+
+const Wrapper = styled(HStack)`
+  position: relative;
+  max-width: 100%;
+`;
 const CollectionNameText = styled(BreadcrumbText)`
   overflow: hidden;
   text-overflow: ellipsis;
@@ -156,4 +154,11 @@ const MobileUsernameText = styled(Paragraph)`
   letter-spacing: -0.04em;
 
   color: ${colors.shadow};
+`;
+
+const SlashText = styled(Paragraph)`
+  font-family: ${TITLE_FONT_FAMILY};
+  font-size: 18px;
+  font-weight: 300;
+  line-height: 21px;
 `;
